+++ CHANGELOG.md
@@ -1,9 +1,10 @@
-WP Plugin Counter Changelog
-===========================
+Slow Website Alert Changelog
+============================
 
 Version 1.0.2 - pending
 
 - Response time is now the average of three `curl` tests.
+- Now includes timestamp in Terminal Notifier alerts.
 
 Version 1.0.1 - 2014-12-04